comment on EVERYTHING

+++ Library/Services/BookCopyService.cs
@@ -10,42 +10,77 @@ using System.Threading.Tasks;
 
 namespace Library.Services
 {
+    /// <summary>
+    /// Class responsible for operations triggered in the GUI about Book Copies.
+    /// </summary>
     class BookCopyService : IService
     {
-        BookCopyRepository bookCopyRepository;
+        private BookCopyRepository bookCopyRepository;
 
+        /// <summary>
+        /// Event representing a change in the current book copy collection.
+        /// </summary>
         public event EventHandler Updated;
-
-        /// <param name="rFactory">A repository factory, so the service can create its own repository.</param>
+        
+        /// <summary>
+        /// Creates and assigns a repository for Book Copies.
+        /// </summary>
+        /// <param name="rFactory"></param>
         public BookCopyService(RepositoryFactory rFactory)
         {
             this.bookCopyRepository = rFactory.CreateBookCopyRepository();
         }
 
+        /// <summary>
+        /// Returns all book copies.
+        /// </summary>
+        /// <returns></returns>
         public IEnumerable<BookCopy> All()
         {
             return bookCopyRepository.All();
         }
 
+        /// <summary>
+        /// Adds a Book Copy to the database by calling its add function.
+        /// </summary>
+        /// <param name="bc"></param>
         public void Add(BookCopy bc)
         {
             bookCopyRepository.Add(bc);
             OnUpdate();
         }
 
+        /// <summary>
+        /// Removes a Book Copy from the database by calling its remove funciton.
+        /// </summary>
+        /// <param name="bc"></param>
         public void Remove(BookCopy bc)
         {
-            // TODO Can't remove if connected to a loan
-            bookCopyRepository.Remove(bc);
-            OnUpdate();
+            if(bc.State == BookCopy.Status.RETURNED)
+            {
+                bookCopyRepository.Remove(bc);
+                OnUpdate();
+            } else
+            {
+                throw new Exception("Copy is connected to a loan");
+            }
         }
 
+        /// <summary>
+        /// Edits a Book Copy.
+        /// </summary>
+        /// <param name="bc"></param>
         public void Edit(BookCopy bc)
         {
             bookCopyRepository.Edit(bc);
             OnUpdate();
         }
 
+        /// <summary>
+        /// Changes a Book Copys stataus to ON_LOAN if it is available.
+        /// </summary>
+        /// <param name="bc"></param>
+        /// <returns></returns>
         public BookCopy SetLoaned(BookCopy bc)
         {
             if (bc.State == BookCopy.Status.RETURNED)
@@ -59,6 +94,11 @@ namespace Library.Services
             }
         }
 
+        /// <summary>
+        /// Returns a Book Copy with a status of RETURNED.
+        /// </summary>
+        /// <param name="books"></param>
+        /// <returns></returns>
         public List<BookCopy> GetAllAvailableCopies(IEnumerable<Book> books)
         {
             List<BookCopy> availableCopies = new List<BookCopy>();
@@ -75,13 +115,18 @@ namespace Library.Services
             return availableCopies;
         }
 
+        /// <summary>
+        /// Returns a Book Copy on ID.
+        /// </summary>
+        /// <param name="Id"></param>
+        /// <returns></returns>
         public BookCopy GetCopyOnId(int Id)
         {
             return bookCopyRepository.Find(Id);
         }
 
         /// <summary>
-        /// 
+        /// Reuturns all Book Copies sorted ascending on a property.
         /// </summary>
         /// <param name="property"></param>
         /// <returns></returns>
@@ -90,11 +135,18 @@ namespace Library.Services
             return All().OrderBy(BuildQuery(property));
         }
 
+        /// <summary>
+        /// Returns all Book Copies sorted descending on a property.
+        /// </summary>
+        /// <param name="property"></param>
+        /// <returns></returns>
         internal IEnumerable<BookCopy> AllDescendingOnProperty(string property)
         {
             return All().OrderByDescending(BuildQuery(property));
         }
 
+        // Builds a query on a property name. Only works with integer properties.
+        // Credit to Balazs Tihanyi on https://stackoverflow.com/questions/9505189/dynamically-generate-linq-queries for inspiration.
         private Func<BookCopy, int> BuildQuery(string property)
         {
             var x = Expression.Parameter(typeof(BookCopy), "x");
@@ -102,6 +154,9 @@ namespace Library.Services
             return Expression.Lambda<Func<BookCopy, int>>(body, x).Compile();
         }
 
+        /// <summary>
+        /// Invokes the updated event.
+        /// </summary>
         private void OnUpdate()
         {
             Updated?.Invoke(this, new EventArgs());
